Prometheus Improvements
