+++ bot/catherinecore.py
@@ -5,13 +5,8 @@ import asyncpg
 import discord
 from aiohttp import ClientSession
 from cogs import EXTENSIONS, VERSION
+from cogs.ext import prometheus
 from discord.ext import commands
-from libs.cog_utils.prometheus_metrics import (
-    Metrics,
-    create_gauges,
-    create_guild_gauges,
-    fill_gauges,
-)
 from libs.ui.pronouns import ApprovePronounsExampleView
 from libs.utils.config import CatherineConfig
 from libs.utils.reloader import Reloader
@@ -42,7 +37,7 @@ class Catherine(commands.Bot):
             **kwargs,
         )
         self.logger: logging.Logger = logging.getLogger("catherine")
-        self.metrics: Metrics = create_gauges()
+        self.metrics = prometheus.create_gauges(self)
         self.session = session
         self.pool = pool
         self.version = str(VERSION)
@@ -67,13 +62,14 @@ class Catherine(commands.Bot):
         await self.load_extension("jishaku")
 
         if self._prometheus_enabled:
+            await self.load_extension("cogs.ext.prometheus")
             prom_host = self._prometheus.get("host", "127.0.0.1")
             prom_port = self._prometheus.get("port", 6789)
 
             await start_http_server(addr=prom_host, port=prom_port)
             self.logger.info("Prometheus Server started on %s:%s", prom_host, prom_port)
 
-            fill_gauges(self)
+            self.metrics.fill_gauges()
 
         if self._dev_mode:
             self._reloader.start()
@@ -82,7 +78,7 @@ class Catherine(commands.Bot):
         if not hasattr(self, "uptime"):
             self.uptime = discord.utils.utcnow()
         elif self._prometheus_enabled and not hasattr(self, "guild_metrics_created"):
-            self.guild_metrics_created = create_guild_gauges(self)
+            self.guild_metrics_created = self.metrics.create_guild_gauges()
 
         curr_user = None if self.user is None else self.user.name
         self.logger.info(f"{curr_user} is fully ready!")
