--- bot/libs/cog_utils/prometheus_metrics/__init__.py
@@ -1,6 +1,0 @@
-from .structs import Metrics as Metrics
-from .utils import (
-    create_gauges as create_gauges,
-    create_guild_gauges as create_guild_gauges,
-    fill_gauges as fill_gauges,
-)
